work in progress, not necessarily working
The plan here is to redo all the options so that they're fetched using the options() getter method in the core class, and updated after sanitization and stuff in the admin class.

The getter method is started, but no work is yet done on the changes to the admin.

There are also some throwaway changes I'd made to the old options stuff before deciding on this course of action.